+++ aeronet/http/test/connection-state_test.cpp
@@ -10,6 +10,7 @@
 #include <unistd.h>
 
 #include <cerrno>
+#include <chrono>
 #include <coroutine>
 #include <cstddef>
 #include <cstdlib>